fix: компиляция графика цены + подробности ошибок в CI
- PriceFormatter: groupThousands ждёт Long, а cents/100 даёт Int
- PriceChart: у DrawScope.drawPath нет параметра pathEffect — штриховка передаётся
  через Stroke(…, pathEffect = …); с этим аргументом не подходила ни одна перегрузка
- ProductListViewModel: явный тип потока истории, иначе LUB от `flowOf(emptyList())`
  и реального запроса не выводился в один тип
- цены в карточке товара и в истории через общий PriceFormatter (был локальный
  форматтер без разрядов — «4200» против «4 200» под графиком)
- ci: аннотация с контекстом после первой строки «e:», потому что список подходящих
  сигнатур Kotlin пишет следующей строкой без префикса, и в аннотации он терялся

+++ app/src/main/java/com/example/otsled/util/PriceFormatter.kt
@@ -20,7 +20,7 @@ object PriceFormatter {
 
         return buildString {
             if (value < 0) append('-')
-            append(groupThousands(rubles))
+            append(groupThousands(rubles.toLong()))
             if (fraction != 0) {
                 append(',')
                 append(fraction.toString().padStart(2, '0'))
